requirements.txt
PyYAML~=6.0.1
grpcio
grpcio-tools
tabulate~=0.9.0